Added comments to scripts, done some cleanup
