+++ Assets/Main.unity
@@ -38,7 +38,7 @@ RenderSettings:
   m_ReflectionIntensity: 1
   m_CustomReflection: {fileID: 0}
   m_Sun: {fileID: 0}
-  m_IndirectSpecularColor: {r: 0.44657844, g: 0.49641222, b: 0.57481694, a: 1}
+  m_IndirectSpecularColor: {r: 0.44657898, g: 0.4964133, b: 0.5748178, a: 1}
 --- !u!157 &3
 LightmapSettings:
   m_ObjectHideFlags: 0
@@ -176,10 +176,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 287916410}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3289950780fdc6142b4717442e69db2c, type: 3}
+  m_Script: {fileID: 11500000, guid: 6fa2edc17eefbf548bb6d165baa96af5, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  dependancies:
+  dependencies:
   - test
 --- !u!1001 &576188572
 Prefab:
@@ -221,47 +221,12 @@ Prefab:
       value: 7
       objectReference: {fileID: 0}
     - target: {fileID: 1865108389543558, guid: 4cb4ba9278878c94bb2d229b38a05016, type: 2}
-      propertyPath: m_TagString
-      value: PuzzlePainting
-      objectReference: {fileID: 0}
-    - target: {fileID: 1002293755362328, guid: 4cb4ba9278878c94bb2d229b38a05016, type: 2}
-      propertyPath: m_Name
-      value: PaintingFrame
-      objectReference: {fileID: 0}
-    - target: {fileID: 1600731479828138, guid: 4cb4ba9278878c94bb2d229b38a05016, type: 2}
-      propertyPath: m_Name
-      value: PaintingImage
-      objectReference: {fileID: 0}
-    - target: {fileID: 1865108389543558, guid: 4cb4ba9278878c94bb2d229b38a05016, type: 2}
       propertyPath: m_Name
       value: Painting
       objectReference: {fileID: 0}
-    - target: {fileID: 23178002759786140, guid: 4cb4ba9278878c94bb2d229b38a05016,
-        type: 2}
-      propertyPath: m_Materials.Array.data[0]
-      value: 
-      objectReference: {fileID: 2100000, guid: 4e5afa6fb2d093641a6dcda3a4351e22, type: 2}
-    m_RemovedComponents:
-    - {fileID: 114147276328218284, guid: 4cb4ba9278878c94bb2d229b38a05016, type: 2}
+    m_RemovedComponents: []
   m_ParentPrefab: {fileID: 100100000, guid: 4cb4ba9278878c94bb2d229b38a05016, type: 2}
   m_IsPrefabParent: 0
---- !u!1 &576188573 stripped
-GameObject:
-  m_PrefabParentObject: {fileID: 1865108389543558, guid: 4cb4ba9278878c94bb2d229b38a05016,
-    type: 2}
-  m_PrefabInternal: {fileID: 576188572}
---- !u!114 &576188574
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 576188573}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f9f5c35c3950d4048ab10b1d81de97e3, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  sceneName: test
 --- !u!1001 &902197218
 Prefab:
   m_ObjectHideFlags: 0
@@ -323,7 +288,8 @@ Prefab:
       value: 
       objectReference: {fileID: 2100000, guid: 3a5830cd15e74104eb8b3f83bcc8f642, type: 2}
     m_RemovedComponents:
-    - {fileID: 114147276328218284, guid: 4cb4ba9278878c94bb2d229b38a05016, type: 2}
+    - {fileID: 0}
+    - {fileID: 114370949375709358, guid: 4cb4ba9278878c94bb2d229b38a05016, type: 2}
   m_ParentPrefab: {fileID: 100100000, guid: 4cb4ba9278878c94bb2d229b38a05016, type: 2}
   m_IsPrefabParent: 0
 --- !u!1 &972973855
@@ -875,3 +841,4 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   sceneName: test
+  dependencies: []
