+++ Assets/Scripts/DoorController.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 3289950780fdc6142b4717442e69db2c
-timeCreated: 1519975044
+guid: 6fa2edc17eefbf548bb6d165baa96af5
+timeCreated: 1520018502
 licenseType: Free
 MonoImporter:
   externalObjects: {}
